+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v0.1.0
 ### Added
 - First version
+- New power sensors: generation power, total load power, battery power, grid total power
+- Battery state of charge sensor
 ### Changed
 ### Fixed